motecomm extended
The basic mote app compiles now. It does not do any fancy things, but periodically sends garbage over the serial interface and toggels a led if it receives a serial packet. You can try it out:

$ cd $TOSROOT/support/sdk/c/sf
$ ./serialsend /dev/ttyUSB0 115200 18
Sending 12
Note: sync
ack
$ ./seriallisten /dev/ttyUSB0 115200
Note: sync
02 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
02 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
^C

Issuing a wrong baud rate makes the program not quit but appear not to work (no feedback whatsoever) - so watch out.

It is obvious that it is just a basic test to get to know the serial tools. However the packet I was SENDING, was apparently received by the mote (I saw the little green LED srialrx flash) but the SerialReceive event was not triggerd (LED 1 was not toggled - which by coincidence is green as well, but I mean the one you can toggle yourself). Maybe this is because I was sending plain bogus, not sure.