cleaned up + Makefile magic
Moved experimental and auxiliary files to different locations to have a tidier main-project directory.

Also everything in the main directory is intended to be useful (i.e. not a test of some sort) - so everything gets compiled automatically and linked to the final executable. Please keep that in mind when adding new source files.