fixed transmission problem
The problem had to do with sending chunks to fast after each other. I now added a delay of 50 ms between transmissions of _chunks_. This is a HACK! But right now I dont know a better way to fix this. I also verified that the transmission __DOES NOT__ corrupt the data, although I could only test it on one pc (I put one mote on USB0 the other on USB1 and started ./client once normally and once with the newly introduced debug-flag 'notun' which disables setting up the tun/tap interface). It works pretty well: I tested it by PRINTING (not sending) a sum of all bytes in the actual packet when sending it and when received. It was always the same. I might add that it looks pretty cool. However I am unable to test it thoughly, because I cannot check if sending works bidirectionally, because I can only start one ./client in full sending mode (if I start the second tun, no packet is sent to the first). Perhaps we can check this on monday.